+++ CHANGELOG.md
@@ -29,6 +29,7 @@
 * Added Enhanced Auto Crafter
 
 #### Changes
+* Removed all functionality from the old Automated Crafting Chamber
 * Changed item order in guide for the Villager Rune and Nether Goo (All runes are now grouped together)
 
 #### Fixes